requirements.txt
# Step 1: Basic FastAPI setup
fastapi==0.109.0
uvicorn[standard]==0.27.0
# Step 2: File upload support
python-multipart==0.0.6
# Step 3: Image processing
pillow==10.2.0
# Step 4: AWS Bedrock integration
boto3==1.34.34
python-dotenv==1.0.0
# Step 5: Selenium web scraping
selenium==4.17.2
webdriver-manager==4.0.1
requests==2.31.0
